GOTO Final checking

+++ Interfaz_desktop/interfaz_cs/intermodular/views/VistaEmpleados.xaml.cs
@@ -45,44 +45,51 @@ namespace intermodular
                             FontSize = 19
 
                         };
-                        btn.Click += (object send, RoutedEventArgs a) =>
+                        btn.Click += async (object send, RoutedEventArgs a) =>
                         {
-                            empleado = u;
-                            txtAyuda.Visibility = Visibility.Hidden;
-                            imgRolPressed = false;
-                            imgNombre.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
-                            imgEmail.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
-                            imgPassword.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
-                            imgRol.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
-                            imgNombre.Visibility = Visibility.Visible;
-                            imgEmail.Visibility = Visibility.Visible;
-                            imgPassword.Visibility = Visibility.Visible;
-                            imgRol.Visibility = Visibility.Visible;
-                            txtNombre.Tag = empleado.name;
-                            txtEmail.Tag = empleado.email;
-                            txtPass.Tag = empleado.passw;
-                            comboBoxRol.Tag = !empleado.rol.Equals("Admin") ? comboBoxRol.Items[0] : comboBoxRol.Items[1];
-                            txtNombre.Text = empleado.name;
-                            txtEmail.Text = empleado.email;
-                            txtPass.Text = empleado.passw;
-                            comboBoxRol.SelectedItem = !empleado.rol.Equals("Admin") ? comboBoxRol.Items[0] : comboBoxRol.Items[1];
-                            txtNombre.IsEnabled = false;
-                            txtEmail.IsEnabled = false;
-                            txtPass.IsEnabled = false;
-                            comboBoxRol.IsEnabled = false;
-                            btnCancelar.Visibility = Visibility.Hidden;
-                            btnEditarZona.Visibility = Visibility.Hidden;
-                            btnCancelar.IsEnabled = false;
-                            btnEditarZona.IsEnabled = false;
-                            btnEliminarZona.Visibility = Visibility.Visible;
-                            if (btnPressed != null)
+                            try
                             {
-                                btnPressed.Background = Brushes.White;
-                                btnPressed.Foreground = Brushes.Black;
+                                await User.getAdmins(Client.currentClient._id);
+                                empleado = u;
+                                txtAyuda.Visibility = Visibility.Hidden;
+                                imgRolPressed = false;
+                                imgNombre.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
+                                imgEmail.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
+                                imgPassword.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
+                                imgRol.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
+                                imgNombre.Visibility = Visibility.Visible;
+                                imgEmail.Visibility = Visibility.Visible;
+                                imgPassword.Visibility = Visibility.Visible;
+                                imgRol.Visibility = !u._id.Equals(User.usuariosAdmin[0]._id) ? Visibility.Visible : Visibility.Hidden;
+                                txtNombre.Tag = empleado.name;
+                                txtEmail.Tag = empleado.email;
+                                txtPass.Tag = empleado.passw;
+                                comboBoxRol.Tag = !empleado.rol.Equals("Admin") ? comboBoxRol.Items[0] : comboBoxRol.Items[1];
+                                txtNombre.Text = empleado.name;
+                                txtEmail.Text = empleado.email;
+                                txtPass.Text = empleado.passw;
+                                comboBoxRol.SelectedItem = !empleado.rol.Equals("Admin") ? comboBoxRol.Items[0] : comboBoxRol.Items[1];
+                                txtNombre.IsEnabled = false;
+                                txtEmail.IsEnabled = false;
+                                txtPass.IsEnabled = false;
+                                comboBoxRol.IsEnabled = false;
+                                btnCancelar.Visibility = Visibility.Hidden;
+                                btnEditarZona.Visibility = Visibility.Hidden;
+                                btnCancelar.IsEnabled = false;
+                                btnEditarZona.IsEnabled = false;
+                                btnEliminarZona.Visibility = Visibility.Visible;
+                                if (btnPressed != null)
+                                {
+                                    btnPressed.Background = Brushes.White;
+                                    btnPressed.Foreground = Brushes.Black;
+                                }
+                                btnPressed = btn;
+                                btn.Background = (Brush)(new BrushConverter().ConvertFrom("#3b7a7a"));
+                                btn.Foreground = Brushes.White;
+                            }catch(Exception exc)
+                            {
+                                MessageBox.Show("Error de conexion", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
                             }
-                            btnPressed = btn;
-                            btn.Background = (Brush)(new BrushConverter().ConvertFrom("#3b7a7a"));
-                            btn.Foreground = Brushes.White;
                         };
 
                         btn.MouseEnter += (object senderMouseEnter, MouseEventArgs mouseEventArg) => {
@@ -278,35 +285,40 @@ namespace intermodular
 
         private async void btnEditarZona_Click(object sender, RoutedEventArgs e)
         {
-            if(validNombre(txtNombre.Text) && isValidEmail(txtEmail.Text) && isValidPass(txtPass.Text))
-            {
-                empleado.name = txtNombre.Text;
-                empleado.email = txtEmail.Text;
-                empleado.passw = txtPass.Text;
-                if(await User.updateUser(empleado._id, empleado))
+            await User.getAdmins(Client.currentClient._id);
+                if (validNombre(txtNombre.Text) && isValidEmail(txtEmail.Text) && isValidPass(txtPass.Text))
                 {
-                    btnPressed.Content = empleado.name;
-                    txtAyuda.Visibility = Visibility.Hidden;
-                    txtNombre.IsEnabled = false;
-                    txtEmail.IsEnabled = false;
-                    txtPass.IsEnabled = false;
-                    comboBoxRol.IsEnabled = false;
-                    txtNombre.Tag = txtNombre.Text;
-                    txtEmail.Tag = txtEmail.Tag;
-                    txtPass.Tag = txtPass.Text;
-                    comboBoxRol.Tag = comboBoxRol.SelectedItem;
-                    imgRolPressed = false;
-                    imgNombre.Visibility = Visibility.Visible;
-                    imgEmail.Visibility = Visibility.Visible;
-                    imgPassword.Visibility = Visibility.Visible;
-                    imgRol.Visibility = Visibility.Visible;
-                    imgNombre.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
-                    imgEmail.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
-                    imgPassword.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
-                    imgRol.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
-                    btnEditarZona.Visibility = Visibility.Hidden;
-                    btnCancelar.Visibility = Visibility.Hidden;
+                    empleado.name = txtNombre.Text;
+                    empleado.email = txtEmail.Text;
+                    empleado.passw = txtPass.Text;
+                    if (await User.updateUser(empleado._id, empleado))
+                    {
+                        btnPressed.Content = empleado.name;
+                        txtAyuda.Visibility = Visibility.Hidden;
+                        txtNombre.IsEnabled = false;
+                        txtEmail.IsEnabled = false;
+                        txtPass.IsEnabled = false;
+                        comboBoxRol.IsEnabled = false;
+                        txtNombre.Tag = txtNombre.Text;
+                        txtEmail.Tag = txtEmail.Tag;
+                        txtPass.Tag = txtPass.Text;
+                        comboBoxRol.Tag = comboBoxRol.SelectedItem;
+                        imgRolPressed = false;
+                        imgNombre.Visibility = Visibility.Visible;
+                        imgEmail.Visibility = Visibility.Visible;
+                        imgPassword.Visibility = Visibility.Visible;
+                        imgRol.Visibility = Visibility.Visible;
+                        imgNombre.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
+                        imgEmail.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
+                        imgPassword.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
+                        imgRol.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
+                        btnEditarZona.Visibility = Visibility.Hidden;
+                        btnCancelar.Visibility = Visibility.Hidden;
+                    }
                 }
+            else
+            {
+
             }
         }
 
@@ -339,6 +351,11 @@ namespace intermodular
         }
 
         private void btnCancelar_Click(object sender, RoutedEventArgs e)
+        {
+            resetearValores();
+        }
+
+        private void resetearValores()
         {
             imgRolPressed = false;
             btnCancelar.IsEnabled = true;
@@ -453,40 +470,44 @@ namespace intermodular
 
         private async void btnEliminarZona_Click(object sender, RoutedEventArgs e)
         {
-            if(await User.deleteUser(empleado._id))
+            await User.getAdmins(Client.currentClient._id);
+            if (User.usuariosAdmin.Count > 1 || User.usuariosAdmin.Count == 1 && !User.usuariosAdmin[0]._id.Equals(empleado._id))
             {
-                empleado = null;
-                stackEmpleados.Children.Remove(btnPressed);
-                btnPressed = null;
-                imgRolPressed = false;
-                btnEliminarZona.Visibility = Visibility.Collapsed;
-                btnAgregar.Visibility = Visibility.Visible;
-                txtAyuda.Visibility = Visibility.Visible;
-                txtNombre.Text = "";
-                txtEmail.Text = "";
-                txtPass.Text = "";
-                comboBoxRol.Text = "";
-                txtNombre.Tag = "";
-                txtEmail.Tag = "";
-                txtPass.Tag = "";
-                comboBoxRol.Tag = "";
-                txtNombre.IsEnabled = false;
-                txtEmail.IsEnabled = false;
-                txtPass.IsEnabled = false;
-                comboBoxRol.IsEnabled = false;
-                imgNombre.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
-                imgEmail.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
-                imgPassword.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
-                imgRol.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
-                imgNombre.Visibility = Visibility.Visible;
-                imgEmail.Visibility = Visibility.Visible;
-                imgPassword.Visibility = Visibility.Visible;
-                imgRol.Visibility = Visibility.Visible;
-                btnCancelar.Visibility = Visibility.Hidden;
-                btnEditarZona.Visibility = Visibility.Hidden;
-                btnCancelar.IsEnabled = false;
-                btnEditarZona.IsEnabled = false;
+                if (await User.deleteUser(empleado._id))
+                {
+                    empleado = null;
+                    stackEmpleados.Children.Remove(btnPressed);
+                    btnPressed = null;
+                    imgRolPressed = false;
+                    btnEliminarZona.Visibility = Visibility.Collapsed;
+                    btnAgregar.Visibility = Visibility.Visible;
+                    txtAyuda.Visibility = Visibility.Visible;
+                    txtNombre.Text = "";
+                    txtEmail.Text = "";
+                    txtPass.Text = "";
+                    comboBoxRol.Text = "";
+                    txtNombre.Tag = "";
+                    txtEmail.Tag = "";
+                    txtPass.Tag = "";
+                    comboBoxRol.Tag = "";
+                    txtNombre.IsEnabled = false;
+                    txtEmail.IsEnabled = false;
+                    txtPass.IsEnabled = false;
+                    comboBoxRol.IsEnabled = false;
+                    imgNombre.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
+                    imgEmail.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
+                    imgPassword.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
+                    imgRol.Source = (ImageSource)new ImageSourceConverter().ConvertFrom("..\\..\\images\\pencil.png");
+                    imgNombre.Visibility = Visibility.Visible;
+                    imgEmail.Visibility = Visibility.Visible;
+                    imgPassword.Visibility = Visibility.Visible;
+                    imgRol.Visibility = Visibility.Visible;
+                    btnCancelar.Visibility = Visibility.Hidden;
+                    btnEditarZona.Visibility = Visibility.Hidden;
+                    btnCancelar.IsEnabled = false;
+                    btnEditarZona.IsEnabled = false;
 
+                }
             }
         }
         private void btns_MouseEnter(object sender, MouseEventArgs e)
